backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
django-debug-toolbar==4.2.0
django-filter==23.5
requests===2.32.5
python-dateutil>=2.8.2
drf-yasg
Pillow
